Made the JSON doclet work with Gradle
