+++ spring-auto-restdocs-json-doclet/src/main/java/capital/scalable/restdocs/jsondoclet/ClassDocumentation.java
@@ -1,0 +1,84 @@
+package capital.scalable.restdocs.jsondoclet;
+
+import java.io.File;
+import java.io.FileWriter;
+import java.io.IOException;
+import java.util.HashMap;
+import java.util.Map;
+
+import com.sun.javadoc.ClassDoc;
+import com.sun.javadoc.FieldDoc;
+import com.sun.javadoc.MethodDoc;
+
+public final class ClassDocumentation {
+    private String comment = "";
+    private Map<String, String> fields = new HashMap<>();
+    private Map<String, MethodDocumentation> methods = new HashMap<>();
+
+    private ClassDocumentation() {
+        // enforce usage of static factory method
+    }
+
+    public static ClassDocumentation fromClassDoc(ClassDoc classDoc) {
+        ClassDocumentation cd = new ClassDocumentation();
+        cd.setComment(classDoc.commentText());
+        for (FieldDoc fieldDoc : classDoc.fields(false)) {
+            cd.addField(fieldDoc.name(), fieldDoc.commentText());
+        }
+        for (MethodDoc methodDoc : classDoc.methods(false)) {
+            cd.addMethod(methodDoc);
+        }
+        return cd;
+    }
+
+    private void setComment(String comment) {
+        this.comment = JsonUtils.escape(comment);
+    }
+
+    private void addField(String name, String comment) {
+        this.fields.put(name, JsonUtils.escape(comment));
+    }
+
+    private void addMethod(MethodDoc methodDoc) {
+        this.methods.put(methodDoc.name(), MethodDocumentation.fromMethodDoc(methodDoc));
+    }
+
+    public void writeToFile(File file) {
+        try (FileWriter fw = new FileWriter(file)) {
+            fw.append(toJson());
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+
+    private String toJson() {
+        StringBuilder builder = new StringBuilder();
+        builder.append("{\"comment\":\"");
+        builder.append(comment);
+        builder.append("\",\"fields\":{");
+        for (Map.Entry<String, String> e : fields.entrySet()) {
+            builder.append("\"");
+            builder.append(e.getKey());
+            builder.append("\":\"");
+            builder.append(e.getValue());
+            builder.append("\",");
+        }
+        if (builder.charAt(builder.length() - 1) == ',') {
+            builder.deleteCharAt(builder.length() - 1);
+        }
+        builder.append("}");
+        builder.append(",\"methods\":{");
+        for (Map.Entry<String, MethodDocumentation> e : methods.entrySet()) {
+            builder.append("\"");
+            builder.append(e.getKey());
+            builder.append("\": ");
+            builder.append(e.getValue().toJson());
+            builder.append(",");
+        }
+        if (builder.charAt(builder.length() - 1) == ',') {
+            builder.deleteCharAt(builder.length() - 1);
+        }
+        builder.append("}}");
+        return builder.toString();
+    }
+}
