+++ spring-auto-restdocs-json-doclet/src/main/java/capital/scalable/restdocs/jsondoclet/DocletAbortException.java
@@ -1,0 +1,8 @@
+package capital.scalable.restdocs.jsondoclet;
+
+public class DocletAbortException extends RuntimeException {
+
+    public DocletAbortException(String message) {
+        super(message);
+    }
+}
