+++ spring-auto-restdocs-json-doclet/src/main/java/capital/scalable/restdocs/jsondoclet/JsonUtils.java
@@ -1,0 +1,12 @@
+package capital.scalable.restdocs.jsondoclet;
+
+public final class JsonUtils {
+
+    private JsonUtils() {
+        // only static utility functions
+    }
+
+    public static String escape(String text) {
+        return text.replace("\"", "\\\"").replace("\n", "\\n");
+    }
+}
